refactor: consolidate duct-sizer folder
