+++ app/tools/duct-sizer/tests/utils.test.ts
