+++ app/tools/duct-sizer/utils.ts
