+++ src/components/air-duct-sizer/FlowRateInput.tsx
@@ -1,5 +1,5 @@
 import React, { ChangeEvent, useCallback } from 'react';
-import { Units } from '../../tools/air-duct-sizer/logic';
+import { Units } from '../../tools/duct-sizer/logic';
 
 interface FlowRateInputProps {
   flowRate: string | number;
