+++ src/config/toolConfig.json
@@ -1,11 +1,11 @@
 {
   "tools": [
     {
-      "id": "air-duct-sizer",
+      "id": "duct-sizer",
       "name": "Air Duct Sizer",
       "description": "Calculate duct sizes and pressure drops",
       "icon": "duct-icon",
-      "componentPath": "@/tools/air-duct-sizer",
+      "componentPath": "@/tools/duct-sizer",
       "category": "HVAC",
       "version": "1.0.0"
     },
