+++ src/tools/duct-sizer/events.ts
