+++ src/tools/duct-sizer/logic.ts
