+++ src/tools/duct-sizer/rules/smacna.json
